LayoutTests/platform/chromium/fast/events/onunload-not-on-body-expected.txt

frame "<!--framePath //<!--frame2-->-->" - has 1 onunload handler(s)
CONSOLE MESSAGE: line 1: Blocked alert('unload') during unload.
you should only see one unload alert appear.
